+++ components/core/types.tsx
@@ -30,3 +30,21 @@ export enum CustomLinkSize {
   BIG = 'big',
   SMALL = 'small',
 }
+
+export enum BreadcrumbsType {
+  HOME = 'home',
+  SEARCH = 'search',
+  DETAIL = 'detail',
+  FAVORITES = 'favorite',
+}
+
+export type BreadcrumbsData = {
+  text: string;
+  route: string;
+  asRoute: string;
+  type:
+    | BreadcrumbsType.HOME
+    | BreadcrumbsType.SEARCH
+    | BreadcrumbsType.DETAIL
+    | BreadcrumbsType.FAVORITES;
+};
