+++ helpers/types.tsx
@@ -182,23 +182,3 @@ export type RawReview = {
     comments_count: number;
   };
 };
-
-// Breadcrumbs
-
-export enum BreadcrumbsType {
-  HOME = 'home',
-  SEARCH = 'search',
-  DETAIL = 'detail',
-  FAVORITES = 'favorite',
-}
-
-export type BreadcrumbsData = {
-  text: string;
-  route: string;
-  asRoute: string;
-  type:
-    | BreadcrumbsType.HOME
-    | BreadcrumbsType.SEARCH
-    | BreadcrumbsType.DETAIL
-    | BreadcrumbsType.FAVORITES;
-};
