+++ store/statics.tsx
@@ -1,4 +1,4 @@
-import { BreadcrumbsType } from '@helpers/types';
+import { BreadcrumbsType } from '../components/core/types';
 
 export const DUBLIN_ID = 91;
 export const BASE_URL = 'https://developers.zomato.com/api/v2.1/';
